Detect connection timeout in the client
Currently, web-client hangs if there is a timeout because there is no
way to terminate a hanging thread as far as I know. The timeout is
hardcoded to 3 seconds, but we could make it a command line option or
something.

The timeout can be tested using "./web-client www.google.com:81"